refactor(edge): isolate stream progression and cleanup logic
